# tutorials/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ghostmaxxing Backend API",
    "version": "1.3.0",
    "summary": "HTTP API, federation, moderation, feed, and diagnostic surfaces for the Ghostmaxxing backend.",
    "description": "This specification is derived from the Express routes in server.js and routes/*.js. ActivityPub and WebFinger payloads are JSON-LD/JRD surfaces generated through Fedify, so those schemas intentionally allow extension fields."
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:4040",
      "description": "Default local server"
    },
    {
      "url": "https://ghostmaxxing.vecna.eu",
      "description": "Production origin"
    }
  ],
...
  "paths": {
...
    "/api/admin/": {
      "get": {
        "tags": ["Admin"],
        "summary": "Serve the moderation UI",
        "security": [{ "adminBasic": [] }],
        "responses": {
          "200": {
            "description": "Moderation HTML page.",
            "content": {
              "text/html": { "schema": { "type": "string" } }
            }
          },
          "401": { "$ref": "#/components/responses/BasicAuthRequired" }
        }
      }
    },
    "/api/admin/pending": {
      "get": {
        "tags": ["Admin"],
        "summary": "List pending uploads",
        "security": [{ "adminBasic": [] }],
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": { "$ref": "#/components/schemas/UploadKind" },
            "description": "Optional pending queue filter."
          }
        ],
        "responses": {
          "200": {
            "description": "Pending uploads, oldest first.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PendingUploadsResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/BasicAuthRequired" }
        }
      }
    },
    "/api/admin/media/{id}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Preview raw pending media",
        "security": [{ "adminBasic": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/UploadId" }
        ],
        "responses": {
          "200": {
            "description": "Pending raw media file.",
            "content": {
              "video/mp4": { "schema": { "type": "string", "contentEncoding": "base64" } },
              "video/webm": { "schema": { "type": "string", "contentEncoding": "base64" } },
              "video/ogg": { "schema": { "type": "string", "contentEncoding": "base64" } },
              "image/png": { "schema": { "type": "string", "contentEncoding": "base64" } },
              "application/octet-stream": { "schema": { "type": "string", "contentEncoding": "base64" } }
            }
          },
          "401": { "$ref": "#/components/responses/BasicAuthRequired" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/admin/approve/{id}": {
      "post": {
        "tags": ["Admin"],
        "summary": "Approve a pending upload",
        "security": [{ "adminBasic": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/UploadId" }
        ],
        "responses": {
          "200": {
            "description": "Upload was processed, published, and delivery was attempted.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApproveUploadResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/BasicAuthRequired" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/admin/reject/{id}": {
      "post": {
        "tags": ["Admin"],
        "summary": "Reject a pending upload",
        "security": [{ "adminBasic": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/UploadId" }
        ],
        "responses": {
          "200": {
            "description": "Upload was rejected and raw media was deleted.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/BasicAuthRequired" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/admin/compose": {
      "get": {
        "tags": ["Admin"],
        "summary": "Serve the staff composer UI",
        "security": [{ "adminBasic": [] }],
        "responses": {
          "200": {
            "description": "Composer HTML page.",
            "content": {
              "text/html": { "schema": { "type": "string" } }
            }
          },
          "401": { "$ref": "#/components/responses/BasicAuthRequired" }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Publish a staff-authored post directly",
        "security": [{ "adminBasic": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/ComposeRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post was published and delivery was attempted.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ComposeResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/BasicAuthRequired" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/posts/{id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Fetch one post as JSON",
        "parameters": [
          { "$ref": "#/components/parameters/PostId" }
        ],
        "responses": {
          "200": {
            "description": "Public post record.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PostResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/posts/{id}": {
      "get": {
        "tags": ["Posts", "Federation"],
        "summary": "Serve a post as HTML preview or ActivityPub object",
        "parameters": [
          { "$ref": "#/components/parameters/PostId" },
          {
            "name": "Accept",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "description": "Use `application/activity+json` or `application/ld+json` for ActivityPub JSON-LD; any other Accept gets HTML."
          }
        ],
        "responses": {
          "200": {
            "description": "Content-negotiated post representation.",
            "content": {
              "text/html": { "schema": { "type": "string" } },
              "application/activity+json": {
                "schema": { "$ref": "#/components/schemas/ActivityPubObject" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "postBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token from LAB_POST_TOKEN. If LAB_POST_TOKEN is empty, the backend skips this check."
      },
      "adminBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "HTTP Basic Auth using GSTMXX_ADMIN_USER and GSTMXX_ADMIN_PASS."
      }
    },
    "parameters": {
      "PostId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
...
      "UploadId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "BasicAuthRequired": {
        "description": "HTTP Basic Auth is required.",
        "headers": {
          "WWW-Authenticate": {
            "schema": { "type": "string" }
          }
        },
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "InternalError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "Uuid": {
        "type": "string",
        "format": "uuid"
      },
      "Url": {
        "type": "string",
        "format": "uri"
      },
      "IsoDateTime": {
        "type": "string",
        "format": "date-time"
      },
      "RawActor": {
        "type": "string",
        "enum": ["video", "ghostyles", "news", "clipboard"]
      },
...
      "UploadKind": {
        "type": "string",
        "enum": ["video", "clipboard"]
      },
      "MediaClass": {
        "type": "string",
        "enum": ["pictures", "videos"]
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["ok", "message"],
        "properties": {
          "ok": { "const": false },
          "message": { "type": "string" }
        },
        "additionalProperties": true
      },
      "MessageResponse": {
        "type": "object",
        "required": ["ok", "message"],
        "properties": {
          "ok": { "const": true },
          "message": { "type": "string" }
        },
        "additionalProperties": true
      },
      "DeliveryResult": {
        "type": ["object", "null"],
        "properties": {
          "followers": { "type": "integer", "minimum": 0 },
          "skipped": { "type": "boolean" },
          "error": { "type": "string" }
        },
        "additionalProperties": true
      },
      "EchoResult": {
        "type": "object",
        "properties": {
          "handle": { "type": "string" },
          "followers": { "type": "integer", "minimum": 0 },
          "error": { "type": "string" }
        },
        "additionalProperties": true
      },
      "Attachment": {
        "type": "object",
        "required": ["type", "url", "mediaType", "name"],
        "properties": {
          "type": { "type": "string", "enum": ["Image", "Video"] },
          "url": { "$ref": "#/components/schemas/Url" },
          "mediaType": { "type": "string" },
          "name": { "type": "string" },
          "thumbnailUrl": { "$ref": "#/components/schemas/Url" }
        },
        "additionalProperties": true
      },
...
      "Post": {
        "type": "object",
        "required": ["id", "type", "actor", "content", "createdAt", "likes", "url"],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "type": { "type": "string", "default": "Note" },
          "actor": { "type": "string" },
          "content": { "type": "string" },
          "createdAt": { "$ref": "#/components/schemas/IsoDateTime" },
          "media": { "$ref": "#/components/schemas/MediaClass" },
          "likes": { "type": "integer", "minimum": 0 },
          "attachment": { "$ref": "#/components/schemas/Attachment" },
          "quoteUrl": { "$ref": "#/components/schemas/Url" },
          "quoteAuthorizationUrl": { "$ref": "#/components/schemas/Url" },
          "title": { "type": "string" },
          "subtitle": { "type": "string" },
          "url": { "$ref": "#/components/schemas/Url" }
        },
        "additionalProperties": true
      },
      "PostSummary": {
        "type": "object",
        "required": ["id", "url"],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "url": { "$ref": "#/components/schemas/Url" }
        },
        "additionalProperties": true
      },
...
      "PendingUpload": {
        "type": "object",
        "required": ["id", "kind", "consentVersion", "createdAt", "mediaUrl"],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "kind": { "$ref": "#/components/schemas/UploadKind" },
          "consentVersion": { "type": "string" },
          "ghostyleId": { "type": ["string", "null"] },
          "appVersion": { "type": ["string", "null"] },
          "userNote": { "type": ["string", "null"] },
          "createdAt": { "$ref": "#/components/schemas/IsoDateTime" },
          "mediaUrl": { "type": "string" }
        },
        "additionalProperties": false
      },
      "PendingUploadsResponse": {
        "type": "object",
        "required": ["ok", "pending"],
        "properties": {
          "ok": { "const": true },
          "pending": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PendingUpload" }
          }
        },
        "additionalProperties": false
      },
      "ApproveUploadResponse": {
        "type": "object",
        "required": ["ok", "message", "kind", "actor", "post", "media", "publicUrl"],
        "properties": {
          "ok": { "const": true },
          "message": { "type": "string" },
          "kind": { "$ref": "#/components/schemas/UploadKind" },
          "actor": { "type": "string" },
          "post": { "$ref": "#/components/schemas/PostSummary" },
          "media": { "type": "string" },
          "thumbnail": { "type": ["string", "null"] },
          "publicUrl": { "$ref": "#/components/schemas/Url" },
          "moderatedAt": { "$ref": "#/components/schemas/IsoDateTime" },
          "federation": { "$ref": "#/components/schemas/DeliveryResult" },
          "echoes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EchoResult" }
          }
        },
        "additionalProperties": true
      },
      "ComposeRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/RawActor",
            "default": "news"
          },
          "title": { "type": "string", "minLength": 1 },
          "subtitle": { "type": "string" },
          "content": { "type": "string" },
          "altText": { "type": "string" },
          "image": {
            "type": "string",
            "contentEncoding": "base64",
            "description": "Optional PNG or JPEG image, maximum 10 MB."
          }
        },
        "additionalProperties": true
      },
      "ComposeResponse": {
        "type": "object",
        "required": ["ok", "message", "actor", "post"],
        "properties": {
          "ok": { "const": true },
          "message": { "type": "string" },
          "actor": { "$ref": "#/components/schemas/RawActor" },
          "post": { "$ref": "#/components/schemas/PostSummary" },
          "federation": { "$ref": "#/components/schemas/DeliveryResult" },
          "echoes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EchoResult" }
          }
        },
        "additionalProperties": true
      },
...
      "PostResponse": {
        "type": "object",
        "required": ["ok", "post"],
        "properties": {
          "ok": { "const": true },
          "post": { "$ref": "#/components/schemas/Post" }
        },
        "additionalProperties": false
      },
...
      "ActivityPubObject": {
        "type": "object",
        "properties": {
          "@context": {},
          "id": { "$ref": "#/components/schemas/Url" },
          "type": { "type": "string" }
        },
        "additionalProperties": true
      },
...
    }
  }
}